+++ .gitignore
@@ -20,6 +20,7 @@
 .env.development
 .env.test
 .env.production
+/backend/.old
 
 npm-debug.log*
 yarn-debug.log*
