--- backend/src/main-logic/test-data/testMatchesData.json
@@ -1,78 +1,0 @@
-[
-    {
-        "match": {
-            "id": 267800918,
-            "tournament_id": 10879090,
-            "state": "open",
-            "player1_id": 166014671,
-            "player2_id": 166014674,
-            "player1_prereq_match_id": null,
-            "player2_prereq_match_id": null,
-            "player1_is_prereq_match_loser": false,
-            "player2_is_prereq_match_loser": false,
-            "winner_id": null,
-            "loser_id": null,
-            "started_at": "2022-03-04T22:22:46.538-03:00",
-            "created_at": "2022-03-04T22:22:46.479-03:00",
-            "updated_at": "2022-03-04T22:22:46.538-03:00",
-            "identifier": "A",
-            "has_attachment": false,
-            "round": 1,
-            "player1_votes": null,
-            "player2_votes": null,
-            "group_id": null,
-            "attachment_count": null,
-            "scheduled_time": null,
-            "location": null,
-            "underway_at": null,
-            "optional": false,
-            "rushb_id": null,
-            "completed_at": null,
-            "suggested_play_order": null,
-            "forfeited": null,
-            "open_graph_image_file_name": null,
-            "open_graph_image_content_type": null,
-            "open_graph_image_file_size": null,
-            "prerequisite_match_ids_csv": "",
-            "scores_csv": ""
-        }
-    },
-    {
-        "match": {
-            "id": 267800919,
-            "tournament_id": 10879090,
-            "state": "open",
-            "player1_id": 166014672,
-            "player2_id": 166014673,
-            "player1_prereq_match_id": null,
-            "player2_prereq_match_id": null,
-            "player1_is_prereq_match_loser": false,
-            "player2_is_prereq_match_loser": false,
-            "winner_id": null,
-            "loser_id": null,
-            "started_at": "2022-03-04T22:22:46.552-03:00",
-            "created_at": "2022-03-04T22:22:46.486-03:00",
-            "updated_at": "2022-03-04T22:22:46.552-03:00",
-            "identifier": "B",
-            "has_attachment": false,
-            "round": 1,
-            "player1_votes": null,
-            "player2_votes": null,
-            "group_id": null,
-            "attachment_count": null,
-            "scheduled_time": null,
-            "location": null,
-            "underway_at": null,
-            "optional": false,
-            "rushb_id": null,
-            "completed_at": null,
-            "suggested_play_order": null,
-            "forfeited": null,
-            "open_graph_image_file_name": null,
-            "open_graph_image_content_type": null,
-            "open_graph_image_file_size": null,
-            "prerequisite_match_ids_csv": "",
-            "scores_csv": ""
-        }
-    }
-]
