--- backend/src/main-logic/test-data/testParticipantData.json
@@ -1,166 +1,0 @@
-[
-    {
-        "participant": {
-            "id": 166014671,
-            "tournament_id": 10879090,
-            "name": "test",
-            "seed": 1,
-            "active": true,
-            "created_at": "2022-03-04T19:16:01.228-06:00",
-            "updated_at": "2022-03-04T19:16:01.228-06:00",
-            "invite_email": null,
-            "final_rank": null,
-            "misc": null,
-            "icon": null,
-            "on_waiting_list": false,
-            "invitation_id": null,
-            "group_id": null,
-            "checked_in_at": null,
-            "ranked_member_id": null,
-            "custom_field_response": null,
-            "clinch": null,
-            "integration_uids": null,
-            "challonge_username": null,
-            "challonge_email_address_verified": null,
-            "removable": true,
-            "participatable_or_invitation_attached": false,
-            "confirm_remove": true,
-            "invitation_pending": false,
-            "display_name_with_invitation_email_address": "test",
-            "email_hash": null,
-            "username": null,
-            "display_name": "test",
-            "attached_participatable_portrait_url": null,
-            "can_check_in": false,
-            "checked_in": false,
-            "reactivatable": false,
-            "check_in_open": false,
-            "group_player_ids": [],
-            "has_irrelevant_seed": false,
-            "ordinal_seed": "1st"
-        }
-    },
-    {
-        "participant": {
-            "id": 166014672,
-            "tournament_id": 10879090,
-            "name": "test2",
-            "seed": 2,
-            "active": true,
-            "created_at": "2022-03-04T19:16:01.261-06:00",
-            "updated_at": "2022-03-04T19:16:01.261-06:00",
-            "invite_email": null,
-            "final_rank": null,
-            "misc": null,
-            "icon": null,
-            "on_waiting_list": false,
-            "invitation_id": null,
-            "group_id": null,
-            "checked_in_at": null,
-            "ranked_member_id": null,
-            "custom_field_response": null,
-            "clinch": null,
-            "integration_uids": null,
-            "challonge_username": null,
-            "challonge_email_address_verified": null,
-            "removable": true,
-            "participatable_or_invitation_attached": false,
-            "confirm_remove": true,
-            "invitation_pending": false,
-            "display_name_with_invitation_email_address": "test2",
-            "email_hash": null,
-            "username": null,
-            "display_name": "test2",
-            "attached_participatable_portrait_url": null,
-            "can_check_in": false,
-            "checked_in": false,
-            "reactivatable": false,
-            "check_in_open": false,
-            "group_player_ids": [],
-            "has_irrelevant_seed": false,
-            "ordinal_seed": "2nd"
-        }
-    },
-    {
-        "participant": {
-            "id": 166014673,
-            "tournament_id": 10879090,
-            "name": "test3",
-            "seed": 3,
-            "active": true,
-            "created_at": "2022-03-04T19:16:01.302-06:00",
-            "updated_at": "2022-03-04T19:16:01.302-06:00",
-            "invite_email": null,
-            "final_rank": null,
-            "misc": null,
-            "icon": null,
-            "on_waiting_list": false,
-            "invitation_id": null,
-            "group_id": null,
-            "checked_in_at": null,
-            "ranked_member_id": null,
-            "custom_field_response": null,
-            "clinch": null,
-            "integration_uids": null,
-            "challonge_username": null,
-            "challonge_email_address_verified": null,
-            "removable": true,
-            "participatable_or_invitation_attached": false,
-            "confirm_remove": true,
-            "invitation_pending": false,
-            "display_name_with_invitation_email_address": "test3",
-            "email_hash": null,
-            "username": null,
-            "display_name": "test3",
-            "attached_participatable_portrait_url": null,
-            "can_check_in": false,
-            "checked_in": false,
-            "reactivatable": false,
-            "check_in_open": false,
-            "group_player_ids": [],
-            "has_irrelevant_seed": false,
-            "ordinal_seed": "3rd"
-        }
-    },
-    {
-        "participant": {
-            "id": 166014674,
-            "tournament_id": 10879090,
-            "name": "test4",
-            "seed": 4,
-            "active": true,
-            "created_at": "2022-03-04T19:16:01.334-06:00",
-            "updated_at": "2022-03-04T19:16:01.334-06:00",
-            "invite_email": null,
-            "final_rank": null,
-            "misc": null,
-            "icon": null,
-            "on_waiting_list": false,
-            "invitation_id": null,
-            "group_id": null,
-            "checked_in_at": null,
-            "ranked_member_id": null,
-            "custom_field_response": null,
-            "clinch": null,
-            "integration_uids": null,
-            "challonge_username": null,
-            "challonge_email_address_verified": null,
-            "removable": true,
-            "participatable_or_invitation_attached": false,
-            "confirm_remove": true,
-            "invitation_pending": false,
-            "display_name_with_invitation_email_address": "test4",
-            "email_hash": null,
-            "username": null,
-            "display_name": "test4",
-            "attached_participatable_portrait_url": null,
-            "can_check_in": false,
-            "checked_in": false,
-            "reactivatable": false,
-            "check_in_open": false,
-            "group_player_ids": [],
-            "has_irrelevant_seed": false,
-            "ordinal_seed": "4th"
-        }
-    }
-]
