--- backend/src/main-logic/test-data/testTournamentData.json
@@ -1,106 +1,0 @@
-[
-    {
-        "tournament": {
-            "id": 10879090,
-            "name": "test",
-            "url": "qbxp4ox7",
-            "description": "",
-            "tournament_type": "single elimination",
-            "started_at": "2022-03-04T20:22:46.494-05:00",
-            "completed_at": null,
-            "require_score_agreement": false,
-            "notify_users_when_matches_open": true,
-            "created_at": "2022-03-04T20:17:30.975-05:00",
-            "updated_at": "2022-03-04T20:22:46.578-05:00",
-            "state": "underway",
-            "open_signup": false,
-            "notify_users_when_the_tournament_ends": true,
-            "progress_meter": 0,
-            "quick_advance": false,
-            "hold_third_place_match": false,
-            "pts_for_game_win": "0.0",
-            "pts_for_game_tie": "0.0",
-            "pts_for_match_win": "1.0",
-            "pts_for_match_tie": "0.5",
-            "pts_for_bye": "1.0",
-            "swiss_rounds": 0,
-            "private": false,
-            "ranked_by": "match wins",
-            "show_rounds": true,
-            "hide_forum": false,
-            "sequential_pairings": false,
-            "accept_attachments": false,
-            "rr_pts_for_game_win": "0.0",
-            "rr_pts_for_game_tie": "0.0",
-            "rr_pts_for_match_win": "1.0",
-            "rr_pts_for_match_tie": "0.5",
-            "created_by_api": false,
-            "credit_capped": false,
-            "category": null,
-            "hide_seeds": false,
-            "prediction_method": 0,
-            "predictions_opened_at": null,
-            "anonymous_voting": false,
-            "max_predictions_per_user": 1,
-            "signup_cap": null,
-            "game_id": 198501,
-            "participants_count": 4,
-            "group_stages_enabled": false,
-            "allow_participant_match_reporting": true,
-            "teams": false,
-            "check_in_duration": null,
-            "start_at": null,
-            "started_checking_in_at": null,
-            "tie_breaks": [
-                "match wins vs tied",
-                "game wins",
-                "points scored"
-            ],
-            "locked_at": null,
-            "event_id": null,
-            "public_predictions_before_start_time": false,
-            "ranked": false,
-            "grand_finals_modifier": null,
-            "predict_the_losers_bracket": false,
-            "spam": null,
-            "ham": null,
-            "rr_iterations": null,
-            "tournament_registration_id": null,
-            "donation_contest_enabled": null,
-            "mandatory_donation": null,
-            "non_elimination_tournament_data": {
-                "participants_per_match": ""
-            },
-            "auto_assign_stations": null,
-            "only_start_matches_with_stations": null,
-            "registration_fee": "0.0",
-            "registration_type": "free",
-            "split_participants": false,
-            "allowed_regions": [],
-            "show_participant_country": null,
-            "program_id": null,
-            "program_classification_ids_allowed": null,
-            "team_size_range": null,
-            "toxic": null,
-            "use_new_style": null,
-            "optional_display_data": {
-                "show_standings": "1",
-                "show_announcements": true
-            },
-            "processing": false,
-            "oauth_application_id": null,
-            "description_source": "",
-            "subdomain": null,
-            "full_challonge_url": "https://challonge.com/qbxp4ox7",
-            "live_image_url": "https://challonge.com/qbxp4ox7.svg",
-            "sign_up_url": null,
-            "review_before_finalizing": true,
-            "accepting_predictions": false,
-            "participants_locked": true,
-            "game_name": "Guilty Gear -Strive-",
-            "participants_swappable": false,
-            "team_convertable": false,
-            "group_stages_were_started": false
-        }
-    }
-]
