+++ backend/src/routes/controller.go
@@ -2,11 +2,7 @@ package routes
 
 import (
 	"net/http"
-	"sort"
-	"time"
 
-	mainlogic "github.com/MarcBernstein0/match-display/src/main-logic"
-	"github.com/MarcBernstein0/match-display/src/models"
 	"github.com/gin-gonic/gin"
 )
 
@@ -16,67 +12,67 @@ func HealthCheck(c *gin.Context) {
 	})
 }
 
-type Date struct {
-	Date time.Time `form:"date" binding:"required" time_format:"2006-01-02"`
-}
+// type Date struct {
+// 	Date time.Time `form:"date" binding:"required" time_format:"2006-01-02"`
+// }
 
-func MatchesGET(fetchData mainlogic.FetchData) gin.HandlerFunc {
-	fn := func(c *gin.Context) {
-		var date Date
-		matches := make([]models.TournamentMatches, 0)
-		if err := c.BindQuery(&date); err != nil {
-			errResponse := models.ErrorResponse{
-				Message:      "did not fill out required 'date' query field",
-				ErrorMessage: err.Error(),
-			}
-			c.JSON(http.StatusBadRequest, errResponse)
-			return
-		}
-		// fmt.Printf("%+v\n", date)
-		// get date
-		// call tournaments
-		tournaments, err := fetchData.FetchTournaments(date.Date.Format("2006-01-02"))
-		if err != nil {
-			errResponse := models.ErrorResponse{
-				Message:      "failed to get tournament data",
-				ErrorMessage: err.Error(),
-			}
-			c.JSON(http.StatusInternalServerError, errResponse)
-			return
-		} else if len(tournaments) == 0 {
-			// no errors but empty tournaments
-			c.JSON(http.StatusOK, matches)
-			return
-		}
-		// fmt.Printf("tournaments %+v\n", tournaments)
-		// call particiapnts
-		participants, err := fetchData.FetchParticipants(tournaments)
-		if err != nil {
-			errResponse := models.ErrorResponse{
-				Message:      "failed to get participant data",
-				ErrorMessage: err.Error(),
-			}
-			c.JSON(http.StatusInternalServerError, errResponse)
-			return
-		}
-		// fmt.Printf("list of participants %+v\n", participants)
-		// call matches
-		matches, err = fetchData.FetchMatches(participants)
-		if err != nil {
-			errResponse := models.ErrorResponse{
-				Message:      "failed to get match data",
-				ErrorMessage: err.Error(),
-			}
-			c.JSON(http.StatusInternalServerError, errResponse)
-			return
-		}
-		// fmt.Printf("list of matches %+v\n", matches)
-		// return matches
-		sort.SliceStable(matches, func(i, j int) bool {
-			return matches[i].GameName <= matches[j].GameName
-		})
+// func MatchesGET(fetchData mainlogic.FetchData) gin.HandlerFunc {
+// 	fn := func(c *gin.Context) {
+// 		var date Date
+// 		matches := make([]models.TournamentMatches, 0)
+// 		if err := c.BindQuery(&date); err != nil {
+// 			errResponse := models.ErrorResponse{
+// 				Message:      "did not fill out required 'date' query field",
+// 				ErrorMessage: err.Error(),
+// 			}
+// 			c.JSON(http.StatusBadRequest, errResponse)
+// 			return
+// 		}
+// 		// fmt.Printf("%+v\n", date)
+// 		// get date
+// 		// call tournaments
+// 		tournaments, err := fetchData.FetchTournaments(date.Date.Format("2006-01-02"))
+// 		if err != nil {
+// 			errResponse := models.ErrorResponse{
+// 				Message:      "failed to get tournament data",
+// 				ErrorMessage: err.Error(),
+// 			}
+// 			c.JSON(http.StatusInternalServerError, errResponse)
+// 			return
+// 		} else if len(tournaments) == 0 {
+// 			// no errors but empty tournaments
+// 			c.JSON(http.StatusOK, matches)
+// 			return
+// 		}
+// 		// fmt.Printf("tournaments %+v\n", tournaments)
+// 		// call particiapnts
+// 		participants, err := fetchData.FetchParticipants(tournaments)
+// 		if err != nil {
+// 			errResponse := models.ErrorResponse{
+// 				Message:      "failed to get participant data",
+// 				ErrorMessage: err.Error(),
+// 			}
+// 			c.JSON(http.StatusInternalServerError, errResponse)
+// 			return
+// 		}
+// 		// fmt.Printf("list of participants %+v\n", participants)
+// 		// call matches
+// 		matches, err = fetchData.FetchMatches(participants)
+// 		if err != nil {
+// 			errResponse := models.ErrorResponse{
+// 				Message:      "failed to get match data",
+// 				ErrorMessage: err.Error(),
+// 			}
+// 			c.JSON(http.StatusInternalServerError, errResponse)
+// 			return
+// 		}
+// 		// fmt.Printf("list of matches %+v\n", matches)
+// 		// return matches
+// 		sort.SliceStable(matches, func(i, j int) bool {
+// 			return matches[i].GameName <= matches[j].GameName
+// 		})
 
-		c.JSON(http.StatusOK, matches)
-	}
-	return fn
-}
+// 		c.JSON(http.StatusOK, matches)
+// 	}
+// 	return fn
+// }
