+++ backend/src/routes/router-setup.go
@@ -1,16 +1,16 @@
 package routes
 
 import (
-	mainlogic "github.com/MarcBernstein0/match-display/src/main-logic"
+	businesslogic "github.com/MarcBernstein0/match-display/src/business-logic"
 	"github.com/gin-gonic/gin"
 )
 
-func RouteSetup(fetchData mainlogic.FetchData) *gin.Engine {
+func RouteSetup(fetchData businesslogic.FetchData) *gin.Engine {
 	r := gin.Default()
 
 	r.Use(Middleware())
 	r.GET("/health", HealthCheck)
-	r.GET("/matches", MatchesGET(fetchData))
+	// r.GET("/matches", MatchesGET(fetchData))
 
 	return r
 }
